day12 part 2
Not entirely happy with the efficiency here. It's certainly not
terrible, but it's not great, either. This is the first part
this year which has taken more than 5 ms, and it takes a few hundred.

Then again, at least the code is fairly straightforward. Also,
it remained the case that the code produced the correct answer the
first time I ran it, which is always nice.

I just have the strong feeling that there's a more efficient way
to solve this, and I totally missed it.

$ hyperfine 'target/release/day12 --part2 --no-part1' --warmup 5
Sun 12 Dec 2021 05:14:33 PM CET
Benchmark #1: target/release/day12 --part2 --no-part1
  Time (mean ± σ):     267.5 ms ±   4.0 ms    [User: 260.2 ms, System: 7.2 ms]
  Range (min … max):   262.6 ms … 275.3 ms    11 runs