day18: fix bugs causing test failures
The lesson here: all nodes need to live on the heap. You can't just
pull a node off the heap, then throw it back on, just because it
didn't used to work like that, because then the up-links from all
depth 1 nodes break.

Splitting still doesn't work, but this is still a big win!