day18 part 1
Wow, that took a while! Feels good to have it finally working; but
I feel like an idiot for taking so long to realize what the problem
was. Oh well, it's a learning experience I suppose.

Also a learning experience: these benchmarks.

$ hyperfine --warmup 10 './tree18' './list18'
Tue 21 Dec 2021 03:54:41 PM CET
Benchmark #1: ./tree18
  Time (mean ± σ):      13.5 ms ±   0.8 ms    [User: 13.0 ms, System: 0.4 ms]
  Range (min … max):    12.8 ms …  21.0 ms    221 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark #2: ./list18
  Time (mean ± σ):       1.7 ms ±   0.1 ms    [User: 1.6 ms, System: 0.1 ms]
  Range (min … max):     1.4 ms …   2.5 ms    1378 runs

  Warning: Command took less than 5 ms to complete. Results might be inaccurate.

Summary
  './list18' ran
    7.98 ± 0.73 times faster than './tree18'

It's really not clear to me why the tree implementation was so much
slower than the tree version, but at a guess, there's some internal
locking in the `RefCell` implementation which is slowing things down.

Aaaaaaaaaaagh. Oh well.